+++ app-libs/stf/Cargo.toml
@@ -39,8 +39,8 @@ sp-runtime = { default-features = false, git = "https://github.com/paritytech/su
 my-node-runtime = { package = "integritee-node-runtime", optional = true, git = "https://github.com/integritee-network/integritee-node.git", branch = "master" }
 pallet-parentchain = { default-features = false, git = "https://github.com/integritee-network/pallets.git", branch = "master" }
 
-# oli dependencies
-poc = { path = "src/poc" }
+# oli/poc
+poc = {path = "poc"}
 
 [dev-dependencies]
 sp-keyring = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
