+++ app-libs/stf/src/getter.rs
@@ -21,7 +21,7 @@ use itp_stf_interface::ExecuteGetter;
 use itp_stf_primitives::types::{AccountId, KeyPair, Signature};
 use itp_utils::stringify::account_id_to_string;
 use log::*;
-use poc::get_person;
+use poc::main;
 use sp_runtime::traits::Verify;
 use std::prelude::v1::*;
 
@@ -72,7 +72,6 @@ pub enum TrustedGetter {
 	#[cfg(feature = "evm")]
 	evm_account_storages(AccountId, H160, H256),
 	add_num(AccountId),
-	call_get_person(AccountId),
 }
 
 impl TrustedGetter {
@@ -88,7 +87,6 @@ impl TrustedGetter {
 			#[cfg(feature = "evm")]
 			TrustedGetter::evm_account_storages(sender_account, ..) => sender_account,
 			TrustedGetter::add_num(sender_account) => sender_account,
-			TrustedGetter::call_get_person(sender_account) => sender_account,
 		}
 	}
 
@@ -177,8 +175,6 @@ impl ExecuteGetter for Getter {
 
 					Some(sum.encode())
 				},
-
-				TrustedGetter::call_get_person(_who) => get_person(),
 			},
 			Getter::public(g) => match g {
 				PublicGetter::some_value => Some(42u32.encode()),
