--- app-libs/stf/src/poc/src/main.rs
@@ -1,16 +1,0 @@
-use match_test::{from_json_string, to_json_string};
-
-pub fn get_person() {
-    let json_string = r#"
-        {
-            "name": "John Doe",
-            "age": 30
-        }
-    "#;
-
-    let person = from_json_string(json_string);
-    println!("{:?}", person);
-
-    let json = to_json_string(&person);
-    println!("{}", json);
-}
